Factor out BulletPointedList component
