+++ support-frontend/assets/pages/contributions-landing/components/ContributionThankYou/new-flow/ContributionThankYouContinueToAccount.jsx
@@ -2,13 +2,13 @@ import React, { useState } from 'react';
 import { css } from '@emotion/core';
 import { body } from '@guardian/src-foundations/typography';
 import { space } from '@guardian/src-foundations';
-import { brand } from '@guardian/src-foundations/palette';
 import { Button } from '@guardian/src-button';
 import { ButtonLink } from '@guardian/src-link';
 import { SvgArrowRightStraight } from '@guardian/src-icons';
 import ActionContainer from './components/ActionContainer';
 import ActionHeader from './components/ActionHeader';
 import ActionBody from './components/ActionBody';
+import BulletPointedList from './components/BulletPointedList';
 import styles from './styles';
 
 const bodyText = css`
@@ -36,21 +36,6 @@ const expandableContainerExpanded = css`
   max-height: 500px;
 `;
 
-const list = css`
-  & > * + * {
-    margin-top: ${space[4]}px;
-  }
-`;
-
-const listItem = css`
-  display: flex;
-  align-items: flex-start;
-
-  & > * + * {
-    margin-left: ${space[2]}px;
-  }
-`;
-
 const buttonContainer = css`
   margin-top: ${space[6]}px;
 `;
@@ -69,18 +54,6 @@ const SvgPersonWithTick = () => (
       d="M17.4733 16.4666C19.785 16.4666 22.5083 13.8699 22.5083 10.7033C22.5083 7.53661 20.64 5.69995 17.4733 5.69995C14.3067 5.69995 12.47 7.53661 12.47 10.7033C12.47 13.8699 15.415 16.4666 17.4733 16.4666ZM24.7566 18.9999L26.1183 17.6699L27.5116 19.0949L31.6599 14.9466L32.9899 16.3082L27.5116 21.7866L26.1816 20.4249L24.7566 18.9999ZM17.4733 18.9999C18.835 18.9999 20.1333 19.0949 21.305 19.3166C21.6533 22.6099 24.25 25.2699 27.48 25.8399L28.8733 31.0332L27.575 32.2999H7.30836L6.07336 31.0332L8.57502 21.5332L9.87335 20.2349C12.4067 19.3799 14.6867 18.9999 17.4733 18.9999Z"
       fill="#121212"
     />
-  </svg>
-);
-
-const SvgBullet = () => (
-  <svg
-    width="8"
-    height="8"
-    viewBox="0 0 8 8"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <circle cx="4" cy="4" r="4" fill={`${brand[500]}`} />
   </svg>
 );
 
@@ -112,61 +85,35 @@ const ContributionThankYouContinueToAccount = () => {
       <div css={styles.hideAfterDesktop}>
         <div
           css={
-          isExpanded
-            ? expandableContainerExpanded
-            : expandableContainerCollapsed
-        }
+            isExpanded
+              ? expandableContainerExpanded
+              : expandableContainerCollapsed
+          }
         >
           <p>Stay signed in across all your devices, to:</p>
-          <ul css={list}>
-            <li css={listItem}>
-              <div>
-                <SvgBullet />
-              </div>
-              <div>
-              Remove unnecessary messages asking you for financial support
-              </div>
-            </li>
-            <li css={listItem}>
-              <div>
-                <SvgBullet />
-              </div>
-              <div>
-              Let you easily manage your recurring contributions, subscriptions
-              and newsletters in one place
-              </div>
-            </li>
-          </ul>
+          <BulletPointedList
+            items={[
+              'Remove unnecessary messages asking you for financial support',
+              'Let you easily manage your recurring contributions, subscriptions and newsletters in one place',
+            ]}
+          />
           <p>
-          Make sure you sign in on each of the devices you use to read our
-          journalism – either today or next time you use them.
+            Make sure you sign in on each of the devices you use to read our
+            journalism – either today or next time you use them.
           </p>
         </div>
       </div>
       <div css={styles.hideBeforeDesktop}>
         <div css={expandableContainerExpanded}>
-          <ul css={list}>
-            <li css={listItem}>
-              <div>
-                <SvgBullet />
-              </div>
-              <div>
-              Remove unnecessary messages asking you for financial support
-              </div>
-            </li>
-            <li css={listItem}>
-              <div>
-                <SvgBullet />
-              </div>
-              <div>
-              Let you easily manage your recurring contributions, subscriptions
-              and newsletters in one place
-              </div>
-            </li>
-          </ul>
+          <BulletPointedList
+            items={[
+              'Remove unnecessary messages asking you for financial support',
+              'Let you easily manage your recurring contributions, subscriptions and newsletters in one place',
+            ]}
+          />
           <p>
-          Make sure you sign in on each of the devices you use to read our
-          journalism – either today or next time you use them.
+            Make sure you sign in on each of the devices you use to read our
+            journalism – either today or next time you use them.
           </p>
         </div>
       </div>
